requirements.txt
Pillow
click==7.1.2
dominate==2.5.2
Flask==1.1.2
Flask-Bootstrap==3.3.7.1
Flask-WTF==0.14.3
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
visitor==0.1.3
Werkzeug==1.0.1
WTForms==2.3.3
SQLAlchemy==1.3.19
Flask-SQLAlchemy==2.4.4
Flask-Login==0.5.0